+++ inst/examples/examples-prep_data_stan.R
@@ -1,0 +1,5 @@
+set.seed(1)
+raw_data <-
+  serocalculator::typhoid_curves_nostrat_100 |>
+  sim_case_data(n = 5)
+prepped_data <- prep_data_stan(raw_data)